move python version to infra
