+++ nomad/tests/test_infras.py
@@ -5,6 +5,7 @@ Test cases for Infra class
 # Imports
 import pytest
 from pathlib import Path
+import requests
 
 
 # Internal imports
@@ -62,7 +63,8 @@ def test_normal_ec2_infra():
     expected_conf = {
         "type": "ec2",
         "instance_type": "c1.medium",
-        "ami_image": "ami-01c647eace872fc02"
+        "ami_image": "ami-01c647eace872fc02",
+        "python_version": "",
     }
     assert infra.infra_conf == expected_conf
 
@@ -76,3 +78,42 @@ def test_bad_ec2():
         _ = Ec2Infra(infra_conf=conf, nomad_wkdir=INFRA_WKDIR)
     expected_msg = "Unsupported value `t2.abcedfg` for key `instance_type`"
     assert str(cm.value) == expected_msg
+
+
+def test_python_major():
+    """
+    Test Python version when only the `major` version is specified. The `major` version
+    is 2.
+    """
+    conf = ec2_tests.PYTHON_VERSION_MAJOR
+    infra = Ec2Infra(infra_conf=conf, nomad_wkdir=INFRA_WKDIR)
+    assert "2.7.18" == infra.infra_conf["python_version"]
+
+
+def test_python_major_minor():
+    """
+    Test Python version when the `major` and `minor` versions are specified.
+    """
+    conf = ec2_tests.PYTHON_VERSION_MAJOR_MINOR
+    infra = Ec2Infra(infra_conf=conf, nomad_wkdir=INFRA_WKDIR)
+    assert "3.6.15" == infra.infra_conf["python_version"]
+
+
+def test_python_major_minor_micro():
+    """
+    Test Python version when the `major`, `minor`, and `micro` are all specified
+    """
+    conf = ec2_tests.PYTHON_VERSION_MAJOR_MINOR_MICRO
+    infra = Ec2Infra(infra_conf=conf, nomad_wkdir=INFRA_WKDIR)
+    assert "3.11.6" == infra.infra_conf["python_version"]
+
+
+def test_python_bad_version():
+    """
+    Test Python version when the `major`, `minor`, and `micro` are all specified
+    """
+    conf = ec2_tests.BAD_PYTHON_VERSION
+    with pytest.raises(requests.exceptions.HTTPError) as cm:
+        _ = Ec2Infra(infra_conf=conf, nomad_wkdir=INFRA_WKDIR)
+    expected_msg = "404 Client Error: Not Found for url: https://www.python.org/ftp/python/3.11.89/"  # noqa
+    assert expected_msg == str(cm.value)
